Draft 1 of midterm project
Added an array to the images and added random code into it, so it randomizes the images.

Added another array to the list of questions.

Added custom functions to keep them organize

Custom functions have stop signs for wrong and right answers for the quiz.

known issues:
The questions all run together, overlapping other questions.

Stop signs are not in front of the images.

Images becoming too large that it overlaps other images.

Planning:
Plan to add random code to the answers.

Plan to add stages to keep the quiz going.

Will add more variety of images to the quiz.