Draft 2 of midterm project
Changed the title of SS6 to WIP1

Changed the background section to gray instead of black when you click the image.

Added a button to start the quiz over

The questions that were overlapping as been fixed.

Fixed the images that becoming too large and overlap other images.

Stop signs are now in front of the images and not the back.

Cut back a bunch of code thanks to the array; now it's much more neater.

Made the question text red to make it see better.

Added more images. (Dogs!)

Known Issues:
Every time you press the next button, another box shows up and doesn't start it again; you have to press it again to make it work.

Having issues trying to randomize the stop signs.

Planning:
Planning to finish the randomizer for the stop sign.

Planning to add color randomizer for the button to make it obvious.