Final Midterm Project
Fixed the next button that caused another button to show up.

Stop signs now randomize. I also added a code that resets if there is no right answer.

Next button now flicker the colors to make it obvious and it is placed in the middle.

Known Issues:
There can be more than one right answers depending on the randomizer, but it's not gamebreaking.